flask-github-navigator/requirements.txt
dependency-injector
flask
bootstrap-flask
pygithub
pyyaml
